web/class_8/redis/requirements.txt
python-dotenv==1.0.1
redis==5.0.3
redis-lru==0.1.2
